Simplify buildtags and Dockerfiles
