--- internal/crypto/fips/doc.go
@@ -1,6 +1,0 @@
-// Copyright 2023 the Pinniped contributors. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-// Package fips can be imported to enable fipsonly tls mode when compiling with fips_strict.
-// It will also cause cgo to be explicitly imported when compiling with fips_strict.
-package fips
