+++ test/testlib/securetls_preference_fips.go
@@ -1,7 +1,7 @@
 // Copyright 2022-2023 the Pinniped contributors. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build fips_strict
+//go:build boringcrypto
 
 package testlib
